renaming and cleanup oh god
