--- src/app/requirements.txt
@@ -1,13 +1,0 @@
-requests
-dash
-dash-renderer
-dash-core-components
-dash-html-components
-dash-table-experimentS
-plotly
-gunicorn
-pandas
-numpy
-sklearn
-xgboost
-search-google
