+++ src/ufctools/scraping.py
@@ -1,0 +1,14 @@
+from io import StringIO
+import os
+from pathlib import Path
+import pickle
+import re
+import requests
+import sys
+from typing import Dict, List, Tuple
+
+from bs4 import BeautifulSoup
+import pandas as pd
+from tqdm import tqdm
+
+from utils
